% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{HeatmapEVmarkers}
\alias{HeatmapEVmarkers}
\title{Quality control of EV data}
\usage{
HeatmapEVmarkers(
  x,
  Fac = NULL,
  color = NULL,
  plotHeat = TRUE,
  IncludeClinical = TRUE,
  IncludeAlbumine = TRUE,
  fontSizeCol = 12,
  fontSizeRow = 12,
  specie = "Homo sapiens"
)
}
\arguments{
\item{x}{(data frame with a GN column)}

\item{Fac}{(Fac - should have length nrow(DF)-1)}

\item{plotHeat}{(bool)}

\item{IncludeClinical}{(bool)}

\item{IncludeAlbumine}{(bool)}

\item{fontSizeCol}{(int)}

\item{fontSizeRow}{(int)}

\item{specie}{(string)}

\item{color=NULL}{(vector with ten colors)}
}
\value{
list
}
\description{
Quality control of EV data
}
\examples{
---
}
\keyword{Quality_control}
