% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{heatmapEVqualityFac}
\alias{heatmapEVqualityFac}
\title{Plot heatmap with markers from ISEV paper}
\usage{
heatmapEVqualityFac(
  DF,
  Fac = NULL,
  color = NULL,
  CatColL = NULL,
  IncludeAlbumine = TRUE,
  plotCatDescription = TRUE
)
}
\arguments{
\item{DF}{(DF - data frame with a GN column)}

\item{Fac}{(Fac - should have length nrow(DF)-1)}

\item{color}{(color - vector with ten colors fpr type of marker)}

\item{CatColL}{(color - protein category)}

\item{IncludeAlbumine}{(bool )}

\item{plotCatDescription}{(bool )}
}
\value{
list
}
\description{
Plot heatmap with markers from ISEV paper
}
\examples{
---
}
\keyword{Quality_control}
