% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{heatmapEVsubCD}
\alias{heatmapEVsubCD}
\title{Plot heatmap with subpopulation  markers from cd9, cd63 and CD81 pulldown from Kowal et al 2016 paper}
\usage{
heatmapEVsubCD(
  DF,
  Fac = NULL,
  color = NULL,
  CatColL = NULL,
  IncludeAlbumine = TRUE
)
}
\arguments{
\item{DF}{(DF - data frame with a GN column)}

\item{Fac}{(Fac - should have length ncol(DF)-1)}

\item{color}{(color - vector with ten colors for type of marker)}

\item{CatColL}{(color - protein category)}

\item{IncludeAlbumine}{(bool )}
}
\value{
list
}
\description{
Plot heatmap with subpopulation  markers from cd9, cd63 and CD81 pulldown from Kowal et al 2016 paper
}
\examples{
---
}
\keyword{Quality_control}
