% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{ImportMaxQuantProteinGroups}
\alias{ImportMaxQuantProteinGroups}
\title{Import MaxQuant protein group file}
\usage{
ImportMaxQuantProteinGroups(filename_txt, data_s = "iBAQ")
}
\arguments{
\item{filename_txt}{(string)}

\item{data_s}{(string)}
}
\value{
DF
}
\description{
Import MaxQuant protein group file
}
\examples{
---
}
\keyword{import}
