% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{plotdensity}
\alias{plotdensity}
\title{plot density of expression values indicating expression values for top exosome markers}
\usage{
plotdensity(Mat, Th = 1, specie = "human", annotation = "GN", Ydelta = 0, ...)
}
\arguments{
\item{Mat}{(matrix - with tranformation)}

\item{specie}{(string- human)}

\item{annotation}{(string - GN)}

\item{Ydelta}{(float - y displacement of label)}

\item{Th=1}{(float - threshold used for making the density of expression values)}
}
\value{
bool
}
\description{
plot density of expression values indicating expression values for top exosome markers
}
\examples{
---
}
\keyword{plot}
