% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{statFunc}
\alias{statFunc}
\title{Calculate summary stat from DF}
\usage{
statFunc(DF)
}
\arguments{
\item{DF}{(DF - first column is values and second column is grouping vector)}
}
\value{
list
}
\description{
Calculate summary stat from DF
}
\examples{
---
}
\keyword{stat}
