% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVqualityMSpF.R
\name{scatterPlot}
\alias{scatterPlot}
\title{plot scatter of relative contamination and EV marker abundance}
\usage{
scatterPlot(
  x = NULL,
  Name = "new",
  Title = "",
  src = "all",
  palette = "jco",
  jitter_b = TRUE,
  Psize = 3,
  scaleData = TRUE,
  NoDensity = TRUE,
  addRefLines = FALSE
)
}
\arguments{
\item{x}{(DF- quantitative values in columns and last column holds the gene names - GN)}

\item{Name}{(string - name of the new samples)}

\item{Title}{(string - plot title)}

\item{src}{(string - which reference samples to include)}

\item{palette}{(colorL or palette)}

\item{jitter_b}{(boolean)}

\item{Psize}{(float)}

\item{scaleData}{(boolean)}

\item{NoDensity}{(boolean)}

\item{addRefLines}{(boolean)}
}
\value{
ggplot
}
\description{
plot scatter of relative contamination and EV marker abundance
}
\examples{
---
}
\keyword{Quality_control}
